rename progress card types file and add random percentages
