+++ src/components/atoms/ProgressChart/ProgressChart.type.ts
