+++ src/components/molecules/MemberProgressChart/MemberProgressChart.styles.ts
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native';
+const styles = StyleSheet.create({
+  percent: {
+    fontSize: 14,
+  },
+  percentage: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: 'black',
+  },
+});
+
+export default styles;
